move prog bar
